Initial push of avoidance branch. Details below:
The purpose of this branch is to investigate various ways to avoid obstacles.

The current avoidance algorithm is as follows:
1. Swarmie has ultrasound sensors detecting if there is an obstacle within 0.8 meters.
2. If obstacle is within 0.8 meters, the swarmie pivots to the right.
3. If obstacle is no longer detected then the swarmie continues going towards he goal
4. Else swarmie pivots to the right again

We would like to investigate further ideas of avoiding an obstacle. This push is simply to save/backup various variables that I have been tweaking to better understand the capabilities of the swarmie and its current algorithm.