Testing obstacle avoidance left or right
When facing an obstacle the robot will turn the opposite direction of the obstacle. For example if the robot senses that the obstacle is on it's left then the robot will turn right. If the obstacle is on the right then it will turn to the left. This most likely won't help in a corner case however in our situation we will only have one rock in the way. Thus no corner cases.